Move tests for moved logic to the new test class
